refactor: colocate Kimi resume builder extension
