+++ Sources/VaultAgentRegistry+Kimi.swift
@@ -1,5 +1,3 @@
-import CMUXAgentLaunch
-
 extension CmuxVaultAgentRegistration {
     static var builtInKimi: CmuxVaultAgentRegistration {
         CmuxVaultAgentRegistration(
@@ -9,22 +7,6 @@ extension CmuxVaultAgentRegistration {
             sessionIdSource: .argvOption("--resume"),
             resumeCommand: "{{executable}} --resume {{sessionId}}",
             cwd: .preserve
-        )
-    }
-}
-
-extension AgentResumeCommandBuilder {
-    static func kimiBuiltInResumeArguments(
-        customRegistration: CmuxVaultAgentRegistration,
-        sessionId: String,
-        launchCommand: AgentLaunchCommandSnapshot?
-    ) -> [String]? {
-        guard customRegistration == CmuxVaultAgentRegistration.builtInKimi else { return nil }
-        return AgentResumeArgv().builtInKind(
-            kind: "kimi",
-            sessionId: sessionId,
-            executablePath: launchCommand?.executablePath,
-            arguments: launchCommand?.arguments ?? []
         )
     }
 }
